+++ src/ZoomHub/Config.hs
@@ -19,7 +19,9 @@ import Database.PostgreSQL.Simple.Instances ()
 import Network.URI.Instances ()
 import Network.Wai (Middleware)
 import Squeal.PostgreSQL.Session.Pool (Pool)
+import Web.ClientSession (Key)
 import qualified ZoomHub.Config.AWS as AWS
+import qualified ZoomHub.Config.Kinde as Kinde
 import ZoomHub.Config.ProcessContent (ProcessContent (..))
 import ZoomHub.Config.Uploads (Uploads (..))
 import ZoomHub.Log.LogLevel (LogLevel)
@@ -39,7 +41,9 @@ googleAnalyticsMeasurementId = "G-XLBYM4SR3W"
 data Config = Config
   { apiUser :: APIUser,
     aws :: AWS.Config,
+    kinde :: Kinde.Config,
     baseURI :: BaseURI,
+    clientSessionKey :: !Key,
     contentBaseURI :: ContentBaseURI,
     dbConnInfo :: PGS.ConnectInfo,
     dbConnPool :: Pool Connection,
@@ -57,19 +61,21 @@ data Config = Config
     publicPath :: FilePath,
     staticBaseURI :: StaticBaseURI,
     uploads :: Uploads,
-    version :: String
+    version :: Text
   }
 
 instance ToJSON Config where
   toJSON Config {..} =
     object
       [ "baseURI" .= baseURI,
+        "clientSessionKey" .= ("<redacted>" :: Text),
         "contentBaseURI" .= contentBaseURI,
         "dbConnInfo" .= dbConnInfo,
         "dbConnPoolIdleTime" .= dbConnPoolIdleTime,
         "dbConnPoolMaxResourcesPerStripe" .= dbConnPoolMaxResourcesPerStripe,
         "dbConnPoolNumStripes" .= dbConnPoolNumStripes,
         "environment" .= environment,
+        "kinde" .= kinde,
         "logLevel" .= show logLevel,
         "maxUploadSizeMegabytes" .= maxUploadSizeMegabytes,
         "port" .= port,