+++ src/ZoomHub/Web/Page.hs
@@ -9,11 +9,13 @@ module ZoomHub.Web.Page
     Title (..),
     Path (..),
     title,
+    titleShort,
     analyticsScript,
   )
 where
 
 import Control.Monad (forM_)
+import Data.Maybe (fromMaybe)
 import Data.Text (Text)
 import qualified Lucid as H
 import NeatInterpolation (text)
@@ -21,6 +23,9 @@ import qualified ZoomHub.Config as Config
 
 title :: Text
 title = "ZoomHub · Share and view full-resolution images easily"
+
+titleShort :: Text
+titleShort = "ZoomHub"
 
 newtype Title = Title Text
 
@@ -37,7 +42,9 @@ instance H.ToHtml Path where
 data Page m a = Page
   { pageTitle :: Title,
     pageCanonicalPath :: Maybe Path,
-    pageBody :: H.HtmlT m a
+    pageHeadStyles :: Maybe (H.HtmlT m a),
+    pageBody :: H.HtmlT m a,
+    pageBodyClassName :: Maybe Text
   }
 
 layout :: (Monad m) => Page m a -> H.HtmlT m a
@@ -57,7 +64,11 @@ layout Page {..} = do
       H.link_ [H.rel_ "stylesheet", H.type_ "text/css", H.href_ "/styles/global.css"]
 
       analyticsScript
-    H.body_ [H.class_ "h-full bg-black m-0 p-0"] pageBody
+
+      forM_ pageHeadStyles id
+    H.body_
+      [H.class_ (fromMaybe "h-full bg-black m-0 p-0" pageBodyClassName)]
+      pageBody
 
 --- TODO: Improve how we represent analytics code.
 --- TODO: Pass through `G-*` Google Analytics measurement ID.