+++ src/ZoomHub/Web/Page/ExploreRecentContent.hs
@@ -33,6 +33,8 @@ instance H.ToHtml ExploreRecentContent where
       ( Page
           { pageTitle = Title $ "Explore: Recent — " <> Page.title,
             pageCanonicalPath = Nothing,
+            pageHeadStyles = Nothing,
+            pageBodyClassName = Nothing,
             pageBody =
               H.div_ [H.class_ "flex flex-col justify-center max-w-max mx-auto py-6"] do
                 H.h1_ [H.class_ "text-3xl text-white font-bold mb-2"] "Explore: Recent"